requirements.txt
azure-kusto-data==4.5.1
azure-kusto-ingest==4.5.1
pandas==1.5.3
python-dotenv==1.0.1
requests==2.28.1

'''
install all the dependencies with a single command:
pip install -r requirements.txt
'''
